requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
av==14.2.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.5.0
enum34==1.1.10
faster-whisper==1.1.1
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.3.0
google-auth==2.38.0
google-genai==1.5.0
h11==0.14.0
halo==0.0.31
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
log-symbols==0.0.14
MarkupSafe==3.0.2
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
onnxruntime==1.21.0
openwakeword==0.6.0
outcome==1.3.0.post0
packaging==24.2
protobuf==6.30.0
pvporcupine==1.9.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dotenv==1.0.1
pytweening==1.2.0
PyYAML==6.0.2
realtimestt==0.3.98
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
selenium==4.29.0
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
spinners==0.0.24
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
webrtcvad-wheels==2.0.14
websocket-client==1.8.0
websockets==14.2
wsproto==1.2.0
pyautogui
pyperclip
Pillow==11.1.0
google-genai
opencv-python==4.11.0.86
mediapipe==0.10.21
screen_brightness_control==0.24.2
pynput==1.8.0
pycaw==20240210
comtypes==1.4.10
elevenlabs==1.54.0
pyttsx3==2.98
certifi